Change HW09 P1 to use float array

+++ 2021Spring/Assignments/HW09/cluster.h
@@ -14,12 +14,15 @@
 // t - number of threads.
 // "centers" - an array of local center positions; it has length t.
 // "dists" - an array that stores the calculated distances; it has length t.
+// (if you use padding to resolve the issue then the lengths of these arrays can
+// change accordingly)
 
 // Example input: arr = [0, 1, 3, 4, 6, 6, 7, 8], n = 8, t = 2.
 // centers = [2, 6] (this is calculated in task1.cpp).
 // Expected results: dists = [6, 3].
 // 6 = |0-2| + |1-2| + |3-2| + |4-2|; 3 = |6-6| + |6-6| + |7-6| + |8-6|.
 
-void cluster(const size_t n, const size_t t, const int *arr, const int *centers, int *dists);
+void cluster(const size_t n, const size_t t, const float *arr,
+             const float *centers, float *dists);
 
 #endif
